# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ACR - Attack Chain Reconstruction Engine",
    "description": "Attack Chain Reconstruction Engine: ingest telemetry, run rule-based detection, correlate events with multi-signal scoring, reconstruct attack chains, map them to MITRE ATT&CK and evaluate against synthetic ground truth.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/chains": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "List Chains",
        "operationId": "list_chains_api_chains_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "is_attack",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Attack"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host"
            }
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 100,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Get Chain",
        "operationId": "get_chain_api_chains__chain_id__get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/timeline": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Timeline",
        "operationId": "chain_timeline_api_chains__chain_id__timeline_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          },
          {
            "name": "include_inferred",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Include inferred missing steps",
              "default": false,
              "title": "Include Inferred"
            },
            "description": "Include inferred missing steps"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimelineOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/evidence": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Evidence",
        "operationId": "chain_evidence_api_chains__chain_id__evidence_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceBlock"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/process-tree": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Process Tree",
        "operationId": "chain_process_tree_api_chains__chain_id__process_tree_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProcessTreeNode"
                  },
                  "title": "Response Chain Process Tree Api Chains  Chain Id  Process Tree Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/network": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Network",
        "operationId": "chain_network_api_chains__chain_id__network_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkViewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/investigation": {
      "get": {
        "tags": [
          "chains"
        ],
        "summary": "Chain Investigation",
        "operationId": "chain_investigation_api_chains__chain_id__investigation_get",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          },
          {
            "name": "include_inferred",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Inferred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/feedback": {
      "post": {
        "tags": [
          "chains"
        ],
        "summary": "Post Feedback",
        "operationId": "post_feedback_api_chains__chain_id__feedback_post",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/confirm": {
      "post": {
        "tags": [
          "chains"
        ],
        "summary": "Confirm Chain",
        "operationId": "confirm_chain_api_chains__chain_id__confirm_post",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/FeedbackComment"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chains/{chain_id}/dismiss": {
      "post": {
        "tags": [
          "chains"
        ],
        "summary": "Dismiss Chain",
        "operationId": "dismiss_chain_api_chains__chain_id__dismiss_post",
        "parameters": [
          {
            "name": "chain_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chain Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/FeedbackComment"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChainDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChainDetail": {
        "properties": {
          "chain_id": {
            "type": "string",
            "title": "Chain Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "is_attack": {
            "type": "boolean",
            "title": "Is Attack",
            "default": false
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceOut"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskOut"
          },
          "tactics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tactics"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/TechniqueRef"
            },
            "type": "array",
            "title": "Techniques"
          },
          "hosts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Hosts"
          },
          "users": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Users"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "type": "array",
            "title": "Events"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/DetectionOut"
            },
            "type": "array",
            "title": "Detections"
          },
          "entities": {
            "items": {
              "$ref": "#/components/schemas/EntityRefOut"
            },
            "type": "array",
            "title": "Entities"
          },
          "processes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Processes"
          },
          "attack_path": {
            "items": {
              "$ref": "#/components/schemas/AttackStage"
            },
            "type": "array",
            "title": "Attack Path"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Evidence"
          },
          "evidence_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Evidence Summary"
          },
          "possible_missing_steps": {
            "items": {
              "$ref": "#/components/schemas/MissingStep"
            },
            "type": "array",
            "title": "Possible Missing Steps"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "analyst_feedback": {
            "items": {
              "$ref": "#/components/schemas/AnalystFeedbackOut"
            },
            "type": "array",
            "title": "Analyst Feedback"
          }
        },
        "type": "object",
        "required": [
          "chain_id",
          "status",
          "confidence",
          "risk"
        ],
        "title": "ChainDetail"
      },
      "ChainListOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ChainSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ChainListOut"
      },
...
      "EvidenceBlock": {
        "properties": {
          "chain_id": {
            "type": "string",
            "title": "Chain Id"
          },
          "observed": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Observed"
          },
          "inferred": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "type": "array",
            "title": "Inferred"
          },
          "observed_count": {
            "type": "integer",
            "title": "Observed Count"
          },
          "inferred_count": {
            "type": "integer",
            "title": "Inferred Count"
          }
        },
        "type": "object",
        "required": [
          "chain_id",
          "observed",
          "inferred",
          "observed_count",
          "inferred_count"
        ],
        "title": "EvidenceBlock"
      },
...
      "FeedbackComment": {
        "properties": {
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "analyst": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "FeedbackComment",
        "description": "Optional payload for the shorthand confirm/dismiss endpoints."
      },
      "FeedbackRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "analyst": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status"
        ],
        "title": "FeedbackRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationOut": {
        "properties": {
          "chain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chain Id"
          },
          "timeline": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Timeline"
          },
          "process_tree": {
            "items": {
              "$ref": "#/components/schemas/ProcessTreeNode"
            },
            "type": "array",
            "title": "Process Tree"
          },
          "network": {
            "$ref": "#/components/schemas/NetworkViewOut"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "timeline",
          "process_tree",
          "network",
          "summary"
        ],
        "title": "InvestigationOut"
      },
...
      "NetworkViewOut": {
        "properties": {
          "connections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Connections"
          },
          "dns_queries": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Dns Queries"
          },
          "external_destinations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "External Destinations"
          },
          "graph": {
            "additionalProperties": true,
            "type": "object",
            "title": "Graph"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          }
        },
        "type": "object",
        "title": "NetworkViewOut"
      },
      "ProcessTreeNode": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "process_name": {
            "type": "string",
            "title": "Process Name"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "ppid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ppid"
          },
          "parent_process_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Process Name"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "command_line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Command Line"
          },
          "severity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Severity"
          },
          "technique_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/ProcessTreeNode"
            },
            "type": "array",
            "title": "Children"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "process_name"
        ],
        "title": "ProcessTreeNode"
      },
...
      "TimelineOut": {
        "properties": {
          "chain_id": {
            "type": "string",
            "title": "Chain Id"
          },
          "include_inferred": {
            "type": "boolean",
            "title": "Include Inferred",
            "default": false
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/TimelineItem"
            },
            "type": "array",
            "title": "Items"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "inferred_count": {
            "type": "integer",
            "title": "Inferred Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "chain_id",
          "items",
          "event_count"
        ],
        "title": "TimelineOut"
      },
...
    }
  }
}